ARTEMIS-5250 Add sender state to the management metrics for federation
Adds management and metrics to the outgoing AMQP senders created by a
remote peer that has message federation configured from the local broker
including a view for the remote policies that tracks number of sent
messages for all producers. Also adds a top level federation management
view that shows counts for incoming and outgoing messages for all
federation consumer and producers.

Add a more declarative state handling API for initialization, start,
stop and shutdown to federation resources and the broker connection API
to make management resource create and cleanup more reliable and robust
to avoid and leaked management objects when reconnection attempts stop
or broker connections are manually stopped or are stopped by
configuration updates or configuration remove.

+++ artemis-protocols/artemis-amqp-protocol/src/main/java/org/apache/activemq/artemis/protocol/amqp/proton/DefaultSenderController.java
@@ -447,36 +447,39 @@ public class DefaultSenderController implements SenderController {
    }
 
    @Override
-   public void close() throws Exception {
-      Source source = (Source) protonSender.getSource();
-      String sourceAddress = getSourceAddress(source);
+   public void close(boolean remoteClose) throws Exception {
+      // Only delete temporary resources when the remote has triggered the close with a Detach
+      if (remoteClose) {
+         final Source source = (Source) protonSender.getSource();
+         final String sourceAddress = getSourceAddress(source);
 
-      if (source != null && sourceAddress != null && multicast) {
-         SimpleString queueName = SimpleString.of(sourceAddress);
-         QueueQueryResult result = sessionSPI.queueQuery(queueName, routingTypeToUse, false);
-         if (result.isExists() && source.getDynamic()) {
-            sessionSPI.deleteQueue(queueName);
-         } else {
-            if (source.getDurable() == TerminusDurability.NONE && tempQueueName != null && (source.getExpiryPolicy() == TerminusExpiryPolicy.LINK_DETACH || source.getExpiryPolicy() == TerminusExpiryPolicy.SESSION_END)) {
-               sessionSPI.removeTemporaryQueue(tempQueueName);
+         if (source != null && sourceAddress != null && multicast) {
+            SimpleString queueName = SimpleString.of(sourceAddress);
+            QueueQueryResult result = sessionSPI.queueQuery(queueName, routingTypeToUse, false);
+            if (result.isExists() && source.getDynamic()) {
+               sessionSPI.deleteQueue(queueName);
             } else {
-               String pubId = protonSender.getName();
-               if (pubId.contains("|")) {
-                  pubId = pubId.split("\\|")[0];
-               }
-               SimpleString queue = createQueueName(connection.isUseCoreSubscriptionNaming(), clientId, pubId, shared, global, isVolatile);
-               result = sessionSPI.queueQuery(queue, multicast ? RoutingType.MULTICAST : RoutingType.ANYCAST, false);
-               //only delete if it isn't volatile and has no consumers
-               if (result.isExists() && !isVolatile && result.getConsumerCount() == 0) {
-                  sessionSPI.deleteQueue(queue);
+               if (source.getDurable() == TerminusDurability.NONE && tempQueueName != null && (source.getExpiryPolicy() == TerminusExpiryPolicy.LINK_DETACH || source.getExpiryPolicy() == TerminusExpiryPolicy.SESSION_END)) {
+                  sessionSPI.removeTemporaryQueue(tempQueueName);
+               } else {
+                  String pubId = protonSender.getName();
+                  if (pubId.contains("|")) {
+                     pubId = pubId.split("\\|")[0];
+                  }
+                  SimpleString queue = createQueueName(connection.isUseCoreSubscriptionNaming(), clientId, pubId, shared, global, isVolatile);
+                  result = sessionSPI.queueQuery(queue, multicast ? RoutingType.MULTICAST : RoutingType.ANYCAST, false);
+                  //only delete if it isn't volatile and has no consumers
+                  if (result.isExists() && !isVolatile && result.getConsumerCount() == 0) {
+                     sessionSPI.deleteQueue(queue);
+                  }
                }
             }
-         }
-      } else if (source != null && source.getDynamic() && (source.getExpiryPolicy() == TerminusExpiryPolicy.LINK_DETACH || source.getExpiryPolicy() == TerminusExpiryPolicy.SESSION_END)) {
-         try {
-            sessionSPI.removeTemporaryQueue(SimpleString.of(sourceAddress));
-         } catch (Exception e) {
-            // Ignore on close, its temporary anyway and will be removed later
+         } else if (source != null && source.getDynamic() && (source.getExpiryPolicy() == TerminusExpiryPolicy.LINK_DETACH || source.getExpiryPolicy() == TerminusExpiryPolicy.SESSION_END)) {
+            try {
+               sessionSPI.removeTemporaryQueue(SimpleString.of(sourceAddress));
+            } catch (Exception e) {
+               // Ignore on close, its temporary anyway and will be removed later
+            }
          }
       }
    }
